fix: Style cleanup
Changed the logos from jpgs to pngs.
Made the logo in the login page change based on dark/light mode.
Fixed color text so it doesn't become unreadable on certain browsers.
Changed login button's color on login screen.
Attempted to fix patient css so it covers more of the page.